move useScrollRef to @suite-hooks
